// src/constants/enums/updatePost.ts
export enum UpdatePostForm {
  id = "update__post_id",

  autocomplete = "off",

  title_updated_input_id = "title_input_id",
  title_updated_input_name = "title_input",
  title_updated_input_label = "",
  title_updated_input_role = "title-input",
  title_updated_input_arial_label = "title-input",
  title_updated_input_placeholder = "I'm happy",
  title_updated_input_type = "text",

  content_updated_input_id = "content_input_id",
  content_updated_input_name = "content_input",
  content_updated_input_label = "",
  content_updated_input_role = "content-input",
  content_updated_input_arial_label = "content-input",
  content_updated_input_placeholder = "What's on your mind, %user_firsname",
  content_updated_input_type = "text",
}